+++ gsrs-module-substance-example/src/test/resources/application-test.conf
@@ -19,10 +19,10 @@ logging.level.root=error
 #logging.level.example.substance.validation=trace
 #logging.level.example.imports=trace
 #logging.level.gsrs.module.substance.importers=trace
-#logging.level.gsrs.data_exchange.model=trace
+#logging.level.gsrs.dataexchange.model=trace
 #logging.level.gsrs.controller=trace
 #logging.level.example.imports.readers=trace
-#logging.level.gsrs.dataexchange.processing_actions=trace
+#logging.level.gsrs.dataexchange.processingactions=trace
 
 
 #logging.level.gsrs=TRACE
@@ -59,7 +59,7 @@ gsrs.importAdapterFactories.substances =
 			"adapterName": "SDF Adapter",
 			"importAdapterFactoryClass": "gsrs.module.substance.importers.SDFImportAdapterFactory",
 			"stagingAreaServiceClass": "gsrs.stagingarea.service.DefaultStagingAreaService",
-			"entityServiceClass" :"gsrs.data_exchange.SubstanceStagingAreaEntityService",
+			"entityServiceClass" :"gsrs.dataexchange.SubstanceStagingAreaEntityService",
 			"description" : "SD file importer for general users",
 			"supportedFileExtensions": [
 				"sdf",
@@ -299,7 +299,7 @@ gsrs.importAdapterFactories.substances =
 			"adapterName": "Delimited Text Adapter",
 			"importAdapterFactoryClass": "gsrs.module.substance.importers.DelimTextImportAdapterFactory",
 			"stagingAreaServiceClass": "gsrs.stagingarea.service.DefaultStagingAreaService",
-			"entityServiceClass" :"gsrs.data_exchange.SubstanceStagingAreaEntityService",
+			"entityServiceClass" :"gsrs.dataexchange.SubstanceStagingAreaEntityService",
 			"description" : "text file importer",
 			#extensions belong here because they can override the default set within the class
 			"supportedFileExtensions": [
@@ -432,7 +432,7 @@ gsrs.importAdapterFactories.substances =
 			"adapterName": "Excel File Adapter",
 			"importAdapterFactoryClass": "gsrs.module.substance.importers.ExcelFileImportAdapterFactory",
 			"stagingAreaServiceClass": "gsrs.stagingarea.service.DefaultStagingAreaService",
-			"entityServiceClass" :"gsrs.data_exchange.SubstanceStagingAreaEntityService",
+			"entityServiceClass" :"gsrs.dataexchange.SubstanceStagingAreaEntityService",
 			"description" : "create substances from data in an Excel spreadsheet",
 			#extensions belong here because they can override the default set within the class
 			"supportedFileExtensions": [
@@ -563,7 +563,7 @@ gsrs.importAdapterFactories.substances =
 			"adapterName": "GSRS JSON Adapter",
 			"importAdapterFactoryClass": "gsrs.module.substance.importers.GSRSJSONImportAdapterFactory",
 			"stagingAreaServiceClass": "gsrs.stagingarea.service.DefaultStagingAreaService",
-			"entityServiceClass" :"gsrs.data_exchange.SubstanceStagingAreaEntityService",
+			"entityServiceClass" :"gsrs.dataexchange.SubstanceStagingAreaEntityService",
 			"description" : "GSRS legacy JSON file importer",
 			#extensions belong here because they can override the default set within the class
 			"supportedFileExtensions": [
